DO NOT MERGE add preset home directory which disables the SSHD server
//// ORIGINAL MESSAGE ////
Based on testing data from the Jenkins plugin 'git-plugin' code base, it
appears that the SSHD server can sometimes take an unreasonable amount
of time to setup. Ultimately, this results in the tests timing out after
the default 180 seconds.

It is currently unknown as to a root cause of why the server takes so
long to start. However, the SSHD server is disabled by default for some
time. Unfortunately, the jenkins-test-harness currently defaults to
version 1.651.2, which does not disable the SSHD server.

Switching to a newer version of the jenkins-war does not currently work,
causing test failures.

Since the root cause of failure is slow SSHD startup, setting the port
to disabled *after* we start Jenkins is not acceptable, it would require
waiting for the server to start, and then disable it.

Instead, provide a new PresetData which has the SSHD server
configuration set to disabled.

This isn't really an ideal solution, because it forces users to opt-out
of the SSHD server they probably don't need. However, I couldn't find
a better solution that worked. It might be possible for us to simply
insert the SSHD configuration to the home directory by default, but that
felt a bit too "magical".
//// ORIGINAL MESSAGE ////

Since using @PresetData was difficult, I opted to just force this
version to switch to the no-sshd-server by default as it ran setup(). It
worked to resolve problems in git-plugin but it isn't a suitable
solution to merge.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>